readd Server::getHost

+++ parsing/Settings.cpp
@@ -6,7 +6,7 @@
 /*   By: vilibert <vilibert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/11 12:00:33 by vilibert          #+#    #+#             */
-/*   Updated: 2024/06/13 12:31:09 by vilibert         ###   ########.fr       */
+/*   Updated: 2024/06/13 12:59:05 by vilibert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -130,7 +130,7 @@ void Settings::setup(void)
 	char buffer[INET_ADDRSTRLEN];
 	FD_ZERO(&_read);
 	FD_ZERO(&_write);
-	for(int i = 0; i < _servers.size(); i++)
+	for(int i = 0; i < (int)_servers.size(); i++)
 	{
 		_servers[i].setup();
